add Launcheer back into smart-on-fhir-connector package
