+++ app/components/Header.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { AppBar, Toolbar, Typography, Button } from '@mui/material';
 import { useAuth } from '@stanfordspezi/user-login-register';
 import { useRouter } from 'next/router'
+import { useNavigation } from 'next/navigation';
 
 // TO DO: include currentUser's name next to log out button 
 export default function Header() {
