+++ app/launch/page.tsx
@@ -8,20 +8,8 @@ SPDX-License-Identifier: MIT
    
 */
 
-import { useEffect } from "react";
-import FHIR from "fhirclient";
-import Loading from '../components/Loading';
+import { Launcher } from "@stanfordspezi/smart-on-fhir-connector";
 
-export default function Launcher() {
-    useEffect(() => {
-        FHIR.oauth2.authorize({
-            clientId: process.env.REACT_APP_SMART_CLIENTID,
-            clientSecret: process.env.REACT_APP_SMART_CLIENTSECRET,
-            scope: process.env.REACT_APP_SMART_SCOPE,
-            redirectUri: '/',
-            completeInTarget: true
-        });
-    }, []);
-
-    return <Loading />;
+export default function SMARTonFHIRLauncher() {
+    return <Launcher/>;
 };
